==> analysis/1_Prep_filter_gun_violent_arrests.ipynb <==
This notebook filters the arrest and crimes data from MPD to generate dataframes for use in the `4_Quasi_Experimental_Analysis_Synthetic_Control.ipynb` notebook, because the pre-analysis plan specifies that we analyze the impact of the CGIC 2.0 improvements using 3 different types of arrests:

- The broadest definition of gun-related arrests, which includes both charges categorized as a weapons-violations in which a gun was involved and violent crimes (assault with a dangerous weapon, homicides, and sex abuse) in which a gun was involved
- Arrests for violent crimes (assault with a dangerous weapon, homicides, and sex abuse) in which a gun was used; and
- Arrests for gun-related weapons violations 

This spits out 3 different csvs:
- broad_arrests_161101_190430.csv
- gun_arrests_161101_190430.csv
- violent_gun_arrests_161101_190430.csv

import pandas as pd
import sys
sys.path.append('..')
import os
import re

#Where our data lives
DATA_DIR = os.path.join('..', 'data')

#Read in Data
ARRESTS_FILE = os.path.join(DATA_DIR, 'all_arrests_2016_2019.csv')  #all arrests from 2016.
arrests = pd.read_csv(ARRESTS_FILE, encoding = 'ISO-8859-1', dtype={'CCN':str})

CRIMES_FILE = os.path.join(DATA_DIR, 'DCRCrime_161101_190430.csv')
crimes = pd.read_csv(CRIMES_FILE, dtype={'CCN':str}) 

print(arrests.shape)
print(crimes.shape)

# Violent Gun Arrests

#These are the words that are associated with violent offenses
violent_offenses = ["Homicide", "Assault with a Dangerous Weapon", "Robbery", "Sex Abuse"]
violent_offenses = [x.lower() for x in violent_offenses] #convert to lower case

#Filter data
violent_arrests = arrests[arrests['Arrest Investigation Category/Description'].str.lower().isin(violent_offenses)]
print('Violent Arrests Dimensions', violent_arrests.shape)

# Get the CCNs where the weapons == Guns
dcr_crime_guns_CCNs = crimes[crimes['DCR Weapon'] == 'GUN']['CCN'].copy()

violent_gun_arrests = violent_arrests[violent_arrests.CCN.isin(dcr_crime_guns_CCNs)]

print('Violent Gun Arrests Dimensions', violent_gun_arrests.shape) 

# Gun Arrests

#For current data 

# This is the list of gun-related words and not-gun related words from Daniel Sebastian @ MPD
gun_words = 'gun|pistol|firearm|rifle|capacity|ammunition|felon in possession|cpwol|alter id marks|unlawful shipment' 
not_gun_words = 'bb|inauguration'


# First, filter our dataset so that the arrest description is specifically for weapons violations
weapons_arrests = \
    arrests[arrests['Arrest Investigation Category/Description'] == 'Weapon Violations']

# Get rid of the rows where we would not be able to filter for gun-related words
weapons_arrests = weapons_arrests[(weapons_arrests['Charge Description'].notnull())]

# Filter for gun words, ignoring the case
gun_arrests = weapons_arrests[weapons_arrests['Charge Description']\
                             .str.contains(gun_words, flags = re.IGNORECASE, regex = True)]

#Take out the ones that are not gun-related words but will pop up based on words above
gun_arrests = gun_arrests[~gun_arrests['Charge Description']\
                          .str.contains(not_gun_words,  flags = re.IGNORECASE, regex = True)]

gun_arrests.shape

# Broad Arrests

#The definititon of broad arrests here is the just violent gun arrests and gun arrests concatenated

broad_arrests = violent_gun_arrests.append(gun_arrests)
broad_arrests.drop_duplicates(inplace = True)
broad_arrests.shape

# Write to CSVs

gun_arrests.to_csv(os.path.join(DATA_DIR, 'gun_arrests_161101_190430.csv'), index = False)
violent_gun_arrests.to_csv(os.path.join(DATA_DIR, 'violent_gun_arrests_161101_190430.csv'), index = False)
broad_arrests.to_csv(os.path.join(DATA_DIR, 'broad_arrests_161101_190430.csv'), index = False)
